pyproject.toml
[project]
name = "csv-upc-omg"
version = "0.1.0"
description = "Taking UPCs, translating them, and making a database of the results"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "click>=8.0.0",
    "httpx>=0.28.1",
]
authors = [
    {name = "Philip James", email = "phildini@phildini.net"},
]
license = {text = "BSD-3-Clause"}
keywords = ["csv", "upc", "data-processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
csv-upc-omg = "csv_upc_omg.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
web = [
    "django>=5.2,<6.0",
    "django-environ>=0.11",
    "djangorestframework>=3.14",
    "django-tables2>=2.6",
    "django-filter>=23",
    "django-htmx>=1.17",
    "psycopg2-binary>=2.9",
    "gunicorn>=21",
    "whitenoise>=6.6",
    "django-debug-toolbar>=4.0",
    "django-tasks>=0.3,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "C4", "PTH"]
ignore = ["S101"]  # Allow assert statements in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src/csv_upc_omg --cov-report=term-missing --cov-report=html --cov-fail-under=90"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/test_*"]

[tool.coverage.report]
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.12",
]
